Fix that typo in the usage message
This PR aims to fix a possible typing error in the usage message displayed upon invoking `imgur-screenshot --help`. It does so by replacing the word `isntead` with `instead`, which seems to be what the author of the message meant. (It is spelled correctly in the readme)

This commit is essential for the survival of this project because proper spelling implies a professional disctributed cloud Quality Assurance team, which helps users build confidence in this product.

Thanks!